Make a file storing a fishScript diff for Kettle3D
This is as part of the new launcher we're making, that automatically updates, lets you run different versions, etc. without needing to re-download a file that hasn't changed when updating. This will work by storing information about the downloaded versions of Kettle3D in a file on the users computer.

Also, another feature that'll come with this update is a new Java installer -- so the same download link works on all computers. This makes way for a mobile version of the game

Sorry if I put too much in the commit description.