Please read the commit description.
This is a snapshot of the game in its development process. 

This does NOT contain all the files Kettle3D needs to run, you'll need to copy the modules folder to your Kettle3D data folder.

On Windows, this is in C:\Users\<your name>\AppData\LocalLow\Kettle3D\Kettle3D. AppData is a hidden folder, but you can type it in in the folder address bar at the top of File Explorer.

On Linux, this is in /home/<you>/.config/unity3d/Kettle3D/Kettle3D. Because .config starts with a dot, you'll need to show hidden files. To do this in Nautilus, go to preferences and show hidden files, most other browsers let you right-click on the folder and check this option.

On a Mac, this is in /Users/<you>/Library/Application Support/Kettle3D/Kettle3D. To make things complicating, this is also a hidden folder and I don't know of a way to get into it. Google probably knows.